backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
docker==6.1.3
prometheus-client==0.19.0
pydantic[email]==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
asyncpg==0.29.0
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
docker==7.1.0

